requirements.txt
djangorestframework==3.3.3
Django==1.9.5
psycopg2==2.6.1